+++ duckscript_sdk/src/sdk/std/fs/rm/mod_test.rs
@@ -87,3 +87,28 @@ fn run_path_recursive() {
 
     assert!(!path.exists());
 }
+
+#[test]
+fn run_multiple_paths() {
+    let path1 = Path::new("./target/_duckscript/rm/multiple_paths/file1.txt");
+    assert!(ensure_exists("./target/_duckscript/rm/multiple_paths/file1.txt").is_ok());
+    assert!(path1.exists());
+    let path2 = Path::new("./target/_duckscript/rm/multiple_paths/file2.txt");
+    assert!(ensure_exists("./target/_duckscript/rm/multiple_paths/file2.txt").is_ok());
+    assert!(path2.exists());
+    let path3 = Path::new("./target/_duckscript/rm/multiple_paths/dir/file.txt");
+    assert!(ensure_exists("./target/_duckscript/rm/multiple_paths/dir/file.txt").is_ok());
+    assert!(path3.exists());
+
+    test::run_script_and_validate(
+        vec![create("")],
+        "out = rm -r ./target/_duckscript/rm/multiple_paths/file1.txt ./target/_duckscript/rm/multiple_paths/file2.txt ./target/_duckscript/rm/multiple_paths/dir",
+        CommandValidation::Match("out".to_string(), "true".to_string()),
+    );
+
+    assert!(!path1.exists());
+    assert!(!path2.exists());
+    assert!(!path3.exists());
+
+    fs::remove_dir_all(&Path::new("./target/_duckscript/rm/multiple_paths")).unwrap();
+}