+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ### v0.7.2
 
+* rm command support for multiple paths.
 * Fix: Command `rm` remove file with flag `-r` #158 (thanks @umaYnit)
 * Upgraded rand 0.8